ActivityManager: Change documentation of low RAM
With OC-MR1, for Android Go, we're considering the definition of
"low RAM" to be <= 1GB of RAM.  So we update the definition in
this documentation.  Additionally, we remove the comments about
screen resolution, as various devices might make trade-offs in
terms of modem stack carveout vs. screen resolution, and those
don't concern us as much as just the total amount of RAM.

Bug: 68277942
Test: None.  This is a documentation change.
Change-Id: Ib02024a3e543f8fec11a1735de32ec58278db6e4