Avoid potential failure to boot with bad storage
Discovered by inspection: It is possible in the
event of file system issues to throw an exception
on boot.

This change moves logic previously done in the
PackageStatusStorage constructor to an initialize()
method that can be called on service start up and
handled more gracefully: if there is a storage problem
a warning will be logged and time zone updates
feature will just not be enabled.

Also:
Commit 6e8e023b9653cd52f40583de9915c91b4b84e8b3
switched to FileUtils.createDir() but the mkdir()
(implied by createDir()) wasn't removed along with
it but was redundant.

Units tests run with:

atest FrameworksServiceTestsCases:PackageStatusStorageTest \
    FrameworksServiceTestsCases:PackageTrackerTest

Test: See above
Test: PTS: run pts -m PtsTimeZoneTestCases
Test: Manual test: deleted /data/system/timezone and rebooted, dir
recreated
Change-Id: I84ab3152d0a9d6781e5665e8615b3ab92b6c7a40