Remove @SystemApi from the fields of R.styleable
R.styleable class is marked as @doconly which hide the class from the
API stub libraries. Therefore, it does not make sense to mark a field in
R.styleable class as @SystemApi since it can't be used from any app.

Having this annotation has been harmless though, but it will trigger an
CTS failure when we do the exact match for SystemAPIs; a runtime entity
annotated as @SystemApi have to exist in the documented Api. So removing
the bogus annotations.

Note for searchKeyphrase* attributes and reason for updating
system-current.txt:

The attributes are listed in public.xml *and* marked as @SystemApi in
attrs.xml. The former made it possible for clients to reference these
attributes via 'android:searchKeyphrase*' in their xml files and the
latter made it possible for clients to reference them via
android.R.attr.searchKeyphrase*. However, the former was the only use
case intended (and in fact the latter wasn't used at all.)

Since marking them as @SystemApi has been effectively no-op and will
even cause CTS failure when we do the exact match for SystemAPIs, the
markers are now removed.

Bug: 67891551
Test: m -j checkapi does not report an error

Change-Id: Ica8cdcc7bbb40b4476c39db247d744f5e870cd06